ticket: 5745 tags: pullup
This patch modifies the NIM Kerberos v5 plug-in to use the 
krb5_get_error_message() function to look up the error string 
if the call to krb5_get_init_creds_password() fails. If the call 
to krb5_get_error_message() fails, the caller will failover to 
the previous method of looking up a suitable error message based 
on the error code.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20572 dc483132-0cff-0310-8789-dd5450dbe970