Merge three patches from Novell, with some updates related to work in our tree since they were submitted, and reindenting for one source file that got particularly ugly in the process.
From Savitha R on 2008-02-18 (yes, it was sitting in my queue a while):

Attached patches have fixes for some issues in LDAP backend. 

1. patch-mem-pwd.txt
This patch clears the password in memory, fixes some memory leaks (in
common code) and a segfault in eDirectory specific code.

2. patch-princ-attrs.txt
This patch removes some eDirectory specific attributes from the common
code

3. patch-edir-rights.txt 
This patch is specific to eDirectory code and fixes some rights issue.
It adds and deletes rights to service objects as required on the kerberos 
attributes.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20744 dc483132-0cff-0310-8789-dd5450dbe970