ticket: 6055
UI should lazy init so that init and fini are only called if one
of the UI calls is called.  The problem is that when you call
krb5_get_init_creds_* you don't know if it will call the prompter
or not.  (It won't if the password is saved in the keychain or
if pkinit succeeds.)


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20780 dc483132-0cff-0310-8789-dd5450dbe970