ticket: 6142
Only close windows in response to fini messages from clients.
Only center the windows the first time they are shown for a client.
Improve the math on the NSRect sent to -[NSWindow setFrame:] so dialog windows don't jump around the screen and the title bar stays in the same place.
Refactor repeated view swapping code to -[AuthenticationController swapView:].

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20821 dc483132-0cff-0310-8789-dd5450dbe970