ticket: new subject: KLLastChangedTime should return current time, not 0
cc_context_get_change_time returns 0 the first time it is called
if the CCacheServer is not running.  KLLastChangedTime always 
returned the current time on the first call.  So use globals and 
a mutex to make sure that KLLastChangedTime both monotonically 
increases and also returns the current time on the first call.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20835 dc483132-0cff-0310-8789-dd5450dbe970