ticket: new subhect: netbsd mktemp actually tries to open file
In krb5_def_store_mkey(), mktemp was being invoked with a string WRFILE:....
This returns an error - as it actually tries to open the file.
Move some of the logic that points to the actual filename earlier - so
mktemp works on the .... portion.

Note that the netbsd linker gives a warning on using mktemp as it may be
insecure - but there is no obvious way to avoid it.



git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20896 dc483132-0cff-0310-8789-dd5450dbe970