ticket: new subject: localize format strings, not final error string
errors.c should localize the incoming format string, not
the string produced by vasprintf.  The format string is 
constant and thus can be added to a localization table, 
whereas the output string is not.

Note that this change depends on error_message also 
localizing error table strings (which it does for KfM 
already).


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20927 dc483132-0cff-0310-8789-dd5450dbe970