ticket: 6342 subject: hash db2 code breaks if st_blksize > 64k
The hash db code assumes in places that the block size is no larger
than 64K.  There's a range check in the case where you don't have a
file but provide initialization info.  The btree code will cap the
block size used at 64K.

Apparently Sun's ZFS can report back a block size of 128K, causing the
db2 tests to fail.

Add such a cap to the hash db creation code.

Note that our default configuration is to use the btree code when
creating a new database, so it's unlikely that this will cause
real-world problems unless someone went out of their way to specify
use of the hash format.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@21786 dc483132-0cff-0310-8789-dd5450dbe970