Do not include cookie in outer padata on client
If the cookie is going to be present in the inner padata then
krb5int_fast_process_error is the wrong place to emit it.
Instead it should be added to the padata in the preauth loop.
This patch removes it from the outer padata.
In addition, it is easier if the cookie is stored as a pa_data on the client rather than a krb5_data.

git-svn-id: svn://anonsvn.mit.edu/krb5/branches/fast@22139 dc483132-0cff-0310-8789-dd5450dbe970