ticket: 6496 subject: Fix vector initialization error in KDC preauth code target_version: 1.7 tags: pullup
In the KDC, get_preauth_hint_list had two bugs initializing the
preauth array.  It was allocating 21 extra entries instead of two due
to a typo (harmless), and it was only zeroing up through one extra
entry (harmful).  Adjust the code to use calloc to avoid further
disagreements of this nature.


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@22369 dc483132-0cff-0310-8789-dd5450dbe970