Both in the previous trunk code  and in the fast negotiation code,  referrals and pre-authentication did not work correctly. In particular, the set of preauthentication mechanisms that had been used in the preauth context, the initial set of mechanisms to try, the fast state, and the negotiated state were not reset whenever the realm changed.
Refactor handling of init_creds loop restarts.  There are three cases
where the loop starts up: at init_creds_init time, when negotiation
needs to restart (fast detected or negotiation not supported), or when
referrals cause the realm to change.  Factor out as much code as
possible to be common in these cases.

git-svn-id: svn://anonsvn.mit.edu/krb5/branches/fast-negotiate@23422 dc483132-0cff-0310-8789-dd5450dbe970